+++ example/lib/utils/mock_items.dart
@@ -2,8 +2,9 @@ import 'package:paginated_items_builder_demo/models/post.dart';
 
 class MockItems {
   // ignore: body_might_complete_normally_nullable
-  static T? getByType<T>() {
-    switch (T.toString()) {
+  static dynamic getByType<T>([String? key]) {
+    final typeKey = key ?? T.toString();
+    switch (typeKey) {
       case 'Post':
         return _post as T;
     }